[COMMON] iommu/exynos: fix mapping less than the required with offset
This patch fixes mapping smaller than the amount required.

Example of the problem:
Assume that the following physical chunks are to be mapped.
   1MB@0x55100000, 1MB@0x5F300000, 1MB@0x43700000
A client want to map 1MB from 0x100400 from the start:
   (1MB-1KB)@0x5F300400 + 1KB@0x43700000
But iovmm_map() just maps:
   1MB@0x5F300000
since it does not care about the offset of the physical chunks
in the middle of the scatter-gather list.

Change-Id: Id7d3893a99ee8beefb92b688ab652750ce211fe9
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>