HMP: Modify the runqueue stats to add a new child stat
Originally:

Add a new unweighted load runqueue runnable ratio, made from
the sum of all runnable tasks on the runqueue. The sum is updated
when a task is updated, and tasks are added/removed as they
are enqueued and dequeued.

The resolution of the comparison is reduced so that there are
around 10 bands of busy-ness for each runqueue. The intent is
to match busy-ness more often and choose the least-recently-
disturbed runqueue.

HOWEVER:

This patch is accidentally broken and results in the following
algorithm in hmp_domain_min_load:

* If any CPU is zero, the overall load is zero
* If no CPUs are idle, the domain is 'fully loaded'

Additionally, the time since last migration count is used to
discriminate between idle CPUs. This is a beneficial scheduling
strategy in at least the Quadrant CPU Benchmark.

No other benchmarks have been tested and since this is strictly
speaking a broken patch, their results might be odd.

Change-Id: I7c9a6f39b42d79f667f497510362bf0803efa36a
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Hwang Seungkon <skon.hwang@samsung.com>