sched: hmp: rebalance if a thread needs to be up-migrated
In Exynos5xx0, it is observed that up-migration of thread is
delayed even though load average ratio of the thread is bigger than
up migration threshold.
Up-migration of the thread is occurred during load balance but
currently up migration of the thread is not considered
when next load balance time is determined.
Thus this patch trigger load balance if updated load average ratio
of the thread is bigger than up migration threshold by
setting next load balance time of the runqueue as now.

Change-Id: I0c38580afed6620a81bc4ba884d3bbb987907bcb
Signed-off-by: Changhwan Youn <chaos.youn@samsung.com>