media: scaler: fix blocking function call in isr
sc_device_run() may called in isr of scaler driver but it allocates
memory from ION, IOVMM and so forth if the ratio of down scaling of
the source image is less than 1/4.
This patch moves allocation of intermediate buffers to queue_setup()
from sc_device_run(). The move of the allocation and initialization
of intermediate buffers involves calculation of scaling ratio.

Change-Id: I5789c1920cfabfebb2bdb3d24872dbbbd342b6fb
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>