[COMMON] iommu/exynos: fix prefetch area overlap
If the region specified in a prefetch overlaps with another prefetch
buffer, System MMU falls into unpredictable state.
The problem may happen if System MMU has 6 prefetch buffers and PB_LMM
is 5. Because 6 prefetch buffers are not able to be combined into a
single prefetch buffer, they must be combined into 2 prefetch buffers
even though there is only one region to specify.

The problem arises when there is only one region to speicify to
prefetch buffers and the System MMU has 6 prefetch buffers. One of the
2 combined prefetch buffers remains uninitialized and the System MMU
results unpredictable behavior.

This patch initialize all prefetch buffers if the number of configured
prefetch buffers are more than the number of regions to specify.

Change-Id: I623158ed8357968350af061955c27f8de60a4f7f
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>