sched: HMP: fix potential logical errors
The previous API for hmp_up_migration reset the destination
CPU every time, regardless of if a migration was desired. The code
using it assumed that the value would not be changed unless
a migration was required. In one rare circumstance, this could
have lead to a task migrating to a little CPU at the wrong time.

Fixing that load to a slight logical tweak to make the surrounding
APIs operate a bit more obviously.

Change-Id: I3aa5c56711e69e842e1f388fcc0b906586d69b90
Signed-off-by: Chris Redpath <chris.readpath@arm.com>