[COMMON] iommu/exynos: promote driver registration to arch_initcall
Some device drivers like seiren are registered by subsys_initcall()
that is called earlier than subsys_initcall_sync() and the devices
that are probed before subsys_initcall_sync() are not serviced by
exynos-iommu.

This patch promotes the timing of the exynos-iommu driver registration
to arch_initcall() and arch_initcall_sync().

Change-Id: I99aa317a7ae4b185dd8a0bd741fa46a5da5b9b18
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>