[COMMO] usb: dwc3: move usb register settings
These settings are for both gadget and host,
so the position should be moved to core from gadget.

And the setting for DWC3_GUCTL register was also moved.
because:

For ss bulk-in data packet, when the host detectsa DPP error or
the internal buffer becomes full, it retries with an ACK TP Retry=1.
However, under the followingconditions, the Retry=1 is falsely carried over
to the next burst.
If there is only single active asynchronous SS EP at the time,
the active asynchronous EP is a Bulk IN EP and the burst with the correctly
Retry=1 ACK TP and the next burstbelong to the same transfer.
This patch is the workaround for this bug.

Change-Id: I5c9423becf52ac8ab93ba6e58951f4e7f06727e8
Signed-off-by: Minho Lee <minho@ishtar.(none)>