[COMMON] media: scaler: flush work before destruction of a buffer
A vb2_buffer may queued to a workqueue if it is delivered with a fence
to wait. However, reference to a freed data can be happened if REQBUF
is called during a work is queued to the workqueue. Thus it is required
to flush pending work if the buffer is pushed to the workqueue.

Change-Id: I08c9293839887c62dd3b52486ecc33f258bb1e83
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>