[7420] iommu/exynos: fix NULL pointer to PB master device
master device is used to distinguish pb group in the same sysmmu.
when initializing PB information sysmmu driver should set master
device of each pb group. But it set NULL as PB master device.
because a function that parses master device information returned
NULL not a pointer to master device.
This patch fixes this wrong return value of the function.

Change-Id: I9745b688967e66836b1825f308159b9f03086917
Signed-off-by: Hyunwoong Kim <khw0178.kim@samsung.com>