[COMMON] exynos/iommu: deprecate prop-map child node
'prop-map' child node node in the fattened device tree definition
is just a redundant because the properties in the node is actually
the properties for the parent node, System MMU device.

This patch deprecates 'prop-map' child nodes and the properties in
the node to be the properties of the parent node with some naming
changes.

Change-Id: Id76132031c78f197b4896ce09d655da843620dd5
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>