[COMMON] media: m2m1shot: prevent racing on a single context
m2m1shot is not designed to use the single m2m1shot context by multple
threads because it does not have a queue of user's requests. Thus users
should not deliver a job to m2m1shot device during the same context
is currently prossecing the previous job. All the user's request,
however, should be serialized by a mutex because the driver does not
trust userspace.

Change-Id: I829ef1e3e02d1e361b81179e26f902c69db59984
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>