[COMMON] mmc: Use variable DATA_TIMEOUT from dts
Currently DATA_TIMEOUT for mmc is 0x1FFFFF (max value)
This is ok for emmc. But not ok for slow devices like sdcard.

1) it takes too long to timeout for abnormal or dead sdcard
2) user may wait too long
3) looks like system lock up
4) may cause dpm timeout
5) spec defines 100ms timeout

It is modified to set user-given timeout(ms) provided by dts.

You can specifiy 'data-timeout' in dts like this for mmc0/1/2
==> data-timeout = <100>; /* for 100ms */

If they value isn't given, max timeout (0xffffffff) will be used
as we've done instead.

Change-Id: I6bf9513c191514ccabefa1e75b1621a6685b457b
Signed-off-by: Jaemin Yoo <jmin.yoo@samsung.com>