[COMMON] media: mfc: support the timeout value of interrupt
In case the given stream size is smaller than the frame size
by error, MFC wait the remained stream even if decoding is completed.
We can set the HW timeout value of interrupt in order to prevent this.
HW timeout value depends on the clock cycle.
MFC clock--------timeout value-----------interrupt time
100MHz		100,000,000		1sec
100MHz		50,000,000		500msec
533MHz		533,000,000		1sec
533MHz		266,500,000		500msec
533MHz		200,000,000		375msec

Change-Id: Id69ccb5d1c14cd45f47e89273f3ab5dd12b43a9d
Signed-off-by: Ayoung Sim <a.sim@samsung.com>