[COMMON] iommu/exynos: refector exynos-iommu driver series
Two versions of exynos-iommu driver were there: one is exynos5-iommu
for System MMU version 1, 2 and 3, and the other is exynos7-iommu for
System MMU version 5 and higher. Those System MMUs are very similar
but detailed configurations. Thus most of the driver code and logic
are the same.
This patch gathers most of the code that can be shared between all
System MMU versions and prepares later System MMU versions and
enahancements that can be applied to all versions of System MMUs.

Change-Id: Ic54579e0b2501319d5d75a85ec25a1011d510e72
Signed-off-by: Sungchun Kang <sungchun.kang@samsung.com>