pwm: samsung: apply pwm framwork
Already merged below patch in mainline.

2ad82e805e91ddb465d7679a15a4dfa8048df32c
pwm: samsung: Fix the value of tcnt and tcmp

The TCMPB and TCNTB registers are calculated through the period
and the duty cycle. The basic calculating formula is like below
on PWM invert option:

    TCNTB = period / tin
    TCMPB = period / tin - duty / tin

However, there is a hidden tick at the end of each period on
S3C64XX and EXYNOS series. So, the formula is also changed like
below:

    TCNTB = period / tin - 1
    TCMPB = period / tin - duty / tin - 1

Additionally, PWM does not support duty cycle 0% and 100%. They
are supported with the nearest values as far as possible.

Change-Id: Ib71722f82120202ea163fb6ca8936ec12d59d2a8
Signed-off-by: Changki Kim <changki.kim@samsung.com>