[COMMON] usb: dwc3: Introduce OTG driver for dwc3
This is first release of otg driver for the dwc3 Synopsys USB3 core.
The otg driver implements the otg final state machine and
control the activation of the device controller or host controller.
In this first implementation, only simple DRD mode is implemented,
Currently simplified OTG Final State Machine is used. In future, if
required, complete OTG FSM can be deployed.

Change-Id: I673e8fe86ea2f92f730c0b6473e743ff8bd5316e
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>
Signed-off-by: Kyounghye Yun <k-hye.yun@samsung.com>