[COMMON] media: vb2: Add support for synchronization
To use synchronization, applications call qbuf and set the
V4L2_BUF_FLAG_USE_SYNC flag in the v4l2_buffer. They can then
pass a fence file descriptor in the "reserved" field of v4l2_buffer
or must set it to -1. vb2 imports the acquire fence and stores it in
the vb2_buffer struct so that drivers can wait on it.

A timeline is created for each queue. The timeline is incremented when
the driver returns a buffer. If V4L2_BUF_FLAG_USE_SYNC is set on qbuf, vb2
creates a point on the timeline and returns a release fence file descriptor
for the buffer in the "reserved" field of v4l2_buffer.

Change-Id: Id70c211dc7e34130a14f867d6be5f9d8c9bfa458
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>
Signed-off-by: Benoit Goby <benoit@android.com>