[COMMON] cpuidle: Support status field of idle state
cpuidle framework does not support way to disable idle state at
build time, it is uncomfortable thing in development. This patch
supports "status" to control idle state. If status property is
existed and set to something except "okay" or "ok", idle state
is disactivated.

	CPU_SLEEP_BOOT: cpu-sleep-boot {
		desc = "boot cluster cpu sleep";
		compatible = "exynos,idle-state";
		arm,psci-suspend-param = <0x0010000>;
		entry-latency-us = <35>;
		exit-latency-us = <90>;
		min-residency-us = <750>;
		status = "ok";
	};

If the status property is absent or set to "okay" or "ok", this
idle state is acitvated.

Change-Id: I14813fa551e7c6d0c563e118cc25408ade3bd757
Signed-off-by: Park Bumgyu <bumgyu.park@samsung.com>