[COMMON] media: vb2-ion: remove redundant cache flush
vb2-ion clean or invalidate caches in .map_dmabuf and .unmap_dmabuf
to ensure no stale data is cached in the CPU cache if the buffer is
possibly accessed with the cacheable attribute by the CPU. But it is
better to decide in the driver because the driver knows if the cache
maintenance is required for a specific buffer. Moreover, there is no
chance to maintain the CPU caches if the buffer is passed through
mmap and userptr memory type. Therefore it is guided to maintain CPU
caches in the drivers and all the problems by the CPU caches due to
omitting cache maintenance are responsible for the drivers.

Change-Id: Ief7a1dc6d750b4433f4f01965f2f7e8e4017824f
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>