[COMMON] iommu/exynos: add exynos_iovmm_(un)map_userptr()
exynos_iovmm_map_userptr() allocates iova from IOVMM and map the
physical memory according to the given user into the allocated IO
virtual memory. It also pins the found pages mapped in the userspace
and unpins when they are unmapped from the System MMU page table.

This function should be called in the process context of the given
user address.

Change-Id: Iad71d8c486fd756a32fe102af285e1047f486c9a
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>