[COMMON] media: m2m1shot2: add support for acquire fences
m2m1shot2 accepts 8 source images and a target image for a single
image processing. It also supports acquire and release fences for
all images. To simplify the acquire fence waiter, m2m1shot waits
for acquire fences asynchronously.

m2m1shot2_context.kref is changed to m2m1shot2_context.starter to
count the number of fence waiters and to identify the time to start
image processing when all fences are signaled. m2m1shot2_context.kref
was used to identify when the context is to be freed but it is
actually unneccessary because the context is always freed when
m2m1shot2_release() is called after waiting the pending task.

The counter, m2m1shot2_context.start is initialized even though no
acquire fence is found because, it is always decremented in
m2m1shot2_start_context() to schedule the current context.

Change-Id: I522b5170a8748c6b764c2c3fd79af9570cb06697
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>