[8890] PM: DEVFREQ: change frequency scale function
Previously, the devfreq drivers were using the CAL APIs
for faster processing when frequency scale. However,
if the gating method is using in the bus or top domains,
the synchronization problems can be occur becasue the dfs CAL APIs
are not using synchronize function. So, the scale function
is changed to CCF.

Change-Id: I0e82824d348512781d2a93833feced6acdb79c08
Signed-off-by: Taekki Kim <taekki.kim@samsung.com>