[COMMON] media: vb2-core: do not avoid pinning on the same userptr
videobuf2-core checks if the queued (userptr, size) is the same as the
previously queued (userptr, size) to avoid pinning on the same userptr
repeatedly.
However, if the queued userptr is mmapped, the assumption of the
vb2-core is not true any more because mmap after munmap with the same
size may return the same address with a high probability. Therefore,
vb2-core should assume that the queued userptr is always the new
userptr that has never been queued.

Change-Id: I4fbb529c716c2eb0b979973f0ed26a52623bc7a6
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>