[COMMON] android: sync: move the position of sync_fence.rmnode
It is found that sync_fence.rmnode->next is corrupted by the invalid
reference to of sync_fence.wq by the device drivers that calls
sync_fence_wait() and so forth.

To prevent the corruption, it is moved to the place that is never
modified after the creation of sync_fence.

Change-Id: Iffbb3ba2f5db7b6567d3419801e27eaccbca2e5d
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>