[8890] video: exynos: dsim: Fixed Tte_protect_on timing
When the CMD_lock is low level,
the commands are immediately transferred.
In other words,
when the CMD_lock is high level the commands transfer are locked.

The purpose of te_protect_on is CMD_lock goes to high level
to lock command transferring before get next TE signal.
The te_protect_on timing chagned to 158.

158 means that 15.8ms(16.67ms per 1fps - margin(5%))

Change-Id: I6a7ea78cba3e73e9ab255ef9fce1e211a7cb0732
Signed-off-by: SeWoon Park <seuni.park@samsung.com>