[8890] video: exynos: decon: Disp_factor goes up 5% tuning value
In case that VPP uses rotation and scaling at a same time,
DISP 168MHz clock seems to make under-run condition.
So, it is needed to tune up DISP clock calculation formula to use 336MHz as next
step of 168MHz in case of rotation and scaling.
Disp_factor goes 1 -> 1.05 (5% tuning value for underrun margin).

Change-Id: Ic8c52de2491ff01c304e4ba9b3c521eb79408d49
Signed-off-by: SeWoon Park <seuni.park@samsung.com>