[COMMON] media: mfc: Added ctx_stop_bits to avoid thread conflict.
There were some problems.
1. If a streamoff or release is called when a H/W operation is in progress,
And if the state is HEAD_PARSED or GOT_INST,
MFC driver doesn't wait this operation and clear the buffers which are being used.
This can cause corruption in vb2 structre.
2. The functions which are waiting for the end of H/W operation wake up
if int flags is set in the end of ISR.
But if the next try_run started directly, the waiting functions can not wake up
because the flag are removed in the start of try_run.
This can cause time out in stop_streaming.

To solve above issues, I modified below.
1. Added ctx_stop_bits to prevent other thread's work.
If ctx_stop_bits is set, ctx_work_bits can't set.
This flag is used only stop_streaming and release function.
2. Added wating funtion in stop_streaming and release.
It has no condition about the state and only check hw_lock.
It is to wait for GOT_INST and HEAD_PARSED.

Change-Id: I31a3d2854ecdaec4d94a4f1dba233c364157fe0c
Signed-off-by: Jeonghee Kim <jhhhh.kim@samsung.com>